Temporary branch for patched 6.1.2

+++ com.jaspersoft.studio.data.sql.ui/src/com/jaspersoft/studio/data/sql/text2model/ConvertOrderBy.java
@@ -1,6 +1,14 @@
 /*******************************************************************************
- * Copyright (C) 2010 - 2016. TIBCO Software Inc. 
- * All Rights Reserved. Confidential & Proprietary.
+ * Copyright (C) 2005 - 2014 TIBCO Software Inc. All rights reserved.
+ * http://www.jaspersoft.com.
+ * 
+ * Unless you have purchased  a commercial license agreement from Jaspersoft,
+ * the following license terms  apply:
+ * 
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
  ******************************************************************************/
 package com.jaspersoft.studio.data.sql.text2model;
 
@@ -19,15 +27,17 @@ import com.jaspersoft.studio.data.sql.model.query.orderby.MOrderBy;
 import com.jaspersoft.studio.data.sql.model.query.orderby.MOrderByColumn;
 import com.jaspersoft.studio.data.sql.model.query.orderby.MOrderByExpression;
 import com.jaspersoft.studio.data.sql.model.query.select.MSelect;
-
-import net.sf.jasperreports.eclipse.util.KeyValue;
+import com.jaspersoft.studio.model.util.KeyValue;
 
 public class ConvertOrderBy {
-	public static void convertOrderBy(SQLQueryDesigner designer, OrOrderByColumn cols) {
+	public static void convertOrderBy(SQLQueryDesigner designer,
+			OrOrderByColumn cols) {
 		if (cols == null)
 			return;
 		if (cols instanceof OrderByColumnFull)
-			doColumn(designer, Util.getKeyword(designer.getRoot(), MSelect.class), (OrderByColumnFull) cols);
+			doColumn(designer,
+					Util.getKeyword(designer.getRoot(), MSelect.class),
+					(OrderByColumnFull) cols);
 		else if (cols instanceof OrOrderByColumn) {
 			MSelect msel = Util.getKeyword(designer.getRoot(), MSelect.class);
 			for (OrderByColumnFull fcol : cols.getEntries())
@@ -35,7 +45,8 @@ public class ConvertOrderBy {
 		}
 	}
 
-	private static void doColumn(SQLQueryDesigner designer, MSelect msel, OrderByColumnFull tf) {
+	private static void doColumn(SQLQueryDesigner designer, MSelect msel,
+			OrderByColumnFull tf) {
 		if (tf.getColOrder() != null) {
 			EList<EObject> eContents = tf.getColOrder().eContents();
 			String column = null;
@@ -48,21 +59,24 @@ public class ConvertOrderBy {
 			String table = ConvertUtil.getDbObjectName(eContents, 2);
 			String schema = ConvertUtil.getDbObjectName(eContents, 3);
 			// String catalog = getDbObjectName(eContents, 3);
-			MOrderByColumn mocol = findColumn(msel, schema, table, column, designer);
+			MOrderByColumn mocol = findColumn(msel, schema, table, column,
+					designer);
 			if (mocol != null) {
 				String direction = tf.getDirection();
 				if (direction != null)
-					mocol.setDesc(direction.trim().equalsIgnoreCase(AMKeyword.DESCENDING_KEYWORD.trim()));
+					mocol.setDesc(direction.trim().equalsIgnoreCase(
+							AMKeyword.DESCENDING_KEYWORD.trim()));
 			} else {
-				MOrderByExpression m = new MOrderByExpression(Util.getKeyword(msel.getParent(), MOrderBy.class),
-						column);
+				MOrderByExpression m = new MOrderByExpression(Util.getKeyword(
+						msel.getParent(), MOrderBy.class), column);
 				String direction = tf.getDirection();
 				if (direction != null)
-					m.setDesc(direction.trim().equalsIgnoreCase(AMKeyword.DESCENDING_KEYWORD.trim()));
+					m.setDesc(direction.trim().equalsIgnoreCase(
+							AMKeyword.DESCENDING_KEYWORD.trim()));
 			}
 		} else if (tf.getColOrderInt() > 0)
-			new MOrderByExpression(Util.getKeyword(msel.getParent(), MOrderBy.class),
-					Long.toString(tf.getColOrderInt()));
+			new MOrderByExpression(Util.getKeyword(msel.getParent(),
+					MOrderBy.class), Integer.toString(tf.getColOrderInt()));
 	}
 
 	//
@@ -77,11 +91,14 @@ public class ConvertOrderBy {
 	// return null;
 	// }
 
-	private static MOrderByColumn findColumn(final MSelect msel, final String schema, final String table,
-			final String column, SQLQueryDesigner designer) {
-		KeyValue<MSQLColumn, MFromTable> kv = ConvertUtil.findColumn(msel, schema, table, column, designer);
+	private static MOrderByColumn findColumn(final MSelect msel,
+			final String schema, final String table, final String column,
+			SQLQueryDesigner designer) {
+		KeyValue<MSQLColumn, MFromTable> kv = ConvertUtil.findColumn(msel,
+				schema, table, column, designer);
 		if (kv != null)
-			return new MOrderByColumn(Util.getKeyword(msel.getParent(), MOrderBy.class), kv.key, kv.value);
+			return new MOrderByColumn(Util.getKeyword(msel.getParent(),
+					MOrderBy.class), kv.key, kv.value);
 		return null;
 	}
 }
